chore: optimize the transaction adapter
