+++ frontend/appflowy_flutter/pubspec.yaml
@@ -47,7 +47,7 @@ dependencies:
     # path: /Users/lucas.xu/Desktop/appflowy-editor
     git:
       url: https://github.com/LucasXu0/appflowy-editor.git
-      ref: a9b9928
+      ref: 6e23fa
   appflowy_popover:
     path: packages/appflowy_popover
